dts: bindings: Add properties to Nordic nRF RADIO binding
The existing binding for Nordic's nRF family RADIO peripheral is
limited to nRF SoCs that support BLE DFE. This commit extends
the binding with additional properties that indicate support for
IEEE 802.15.4 protocol and the BLE Direction Finding Extension.
This way, the binding can be used by all Nordic SoCs that have
a radio peripheral.

Signed-off-by: Jedrzej Ciupis <jedrzej.ciupis@nordicsemi.no>